requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anomalib==2.2.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
docstring_parser==0.17.0
einops==0.8.2
fastapi==0.128.0
filelock==3.20.0
fonttools==4.61.1
FrEIA==0.2
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.11
imagecodecs==2026.1.14
ImageIO==2.37.2
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonargparse==4.46.0
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
lightning==2.6.1
lightning-utilities==0.15.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.1
omegaconf==2.3.0
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
packaging==26.0
pandas==3.0.0
pillow==12.1.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytorch-lightning==2.6.1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
rich-argparse==1.7.2
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.1.14
timm==1.0.24
torch==2.10.0+cpu
torchmetrics==1.8.2
torchvision==0.25.0+cpu
tqdm==4.67.3
typer-slim==0.21.1
typeshed_client==2.8.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarl==1.22.0
